fix: refactors repo
