+++ pyproject.toml
@@ -43,6 +43,7 @@ ephys = [
     'wavpack-numcodecs>=0.1.3'
 ]
 imaging = [
+    'argschema',
     'dask-jobqueue',
     'dask-image',
     'bokeh>=2.1.1, <3.0.0',
