+++ scripts/s3_upload.py
@@ -14,8 +14,8 @@ from cluster.config import load_jobqueue_config
 from dask_jobqueue import SLURMCluster
 from distributed import Client
 from s3transfer.constants import GB, MB
-from aind_data_transfer.s3 import S3Uploader
 
+from aind_data_transfer.s3 import S3Uploader
 from aind_data_transfer.util import file_utils
 from aind_data_transfer.util.file_utils import collect_filepaths
