+++ src/aind_data_transfer/jobs/transcode_job.py
@@ -8,10 +8,10 @@ from shutil import copytree, ignore_patterns
 
 from numcodecs import Blosc
 
-from aind_data_transfer.config_loader.configuration_loader import (
+from aind_data_transfer.config_loader.imaging_configuration_loader import (
     ImagingJobConfigurationLoader,
 )
-from aind_data_transfer.readers import ImagingReaders
+from aind_data_transfer.readers.imaging_readers import ImagingReaders
 from aind_data_transfer.util.file_utils import is_cloud_url, parse_cloud_url
 
 LOGGER = logging.getLogger(__name__)
@@ -112,7 +112,9 @@ def _build_ome_zar_cmd(
     return job_cmd
 
 
-def _build_submit_cmd(job_cmd: str, job_configs: dict, wait: bool = False) -> str:
+def _build_submit_cmd(
+    job_cmd: str, job_configs: dict, wait: bool = False
+) -> str:
     submit_args = job_configs["transcode_job"]["submit_args"]
     # FIXME: necessary to wrap job_cmd in quotes
     submit_cmd = f'python {_SUBMIT_SCRIPT} generate-and-launch-run --job_cmd="{job_cmd}"'
@@ -159,7 +161,9 @@ def main():
 
     wait = job_configs["jobs"]["create_ng_link"]
     if wait:
-        LOGGER.info("Will wait for job to terminate before continuing execution")
+        LOGGER.info(
+            "Will wait for job to terminate before continuing execution"
+        )
 
     zarr_out = dest_data_dir + "/" + raw_image_dir_name
     if job_configs["jobs"]["transcode"]:
@@ -179,7 +183,9 @@ def main():
         subprocess.run(ng_link_cmd, shell=True)
         output_json = data_src_dir / "process_output.json"
         if not output_json.is_file():
-            LOGGER.error(f"Creating neuroglancer link failed; {output_json} was not created")
+            LOGGER.error(
+                f"Creating neuroglancer link failed; {output_json} was not created"
+            )
 
     if job_configs["jobs"]["upload_aux_files"]:
         LOGGER.info("Uploading auxiliary data")
