+++ src/aind_data_transfer/readers/__init__.py
@@ -1,0 +1,1 @@
+"""Package for modules to help read data from a various sources."""
