+++ src/aind_data_transfer/transformations/ephys_compressors.py
@@ -13,7 +13,7 @@ from numcodecs import Blosc
 from tqdm import tqdm
 from wavpack_numcodecs import WavPack
 
-from aind_data_transfer.readers import EphysReaders
+from aind_data_transfer.readers.ephys_readers import EphysReaders
 
 
 class EphysCompressors:
@@ -169,33 +169,6 @@ class EphysCompressors:
                     "experiment_name": read_block["experiment_name"],
                     "stream_name": read_block["stream_name"],
                 }
-            )
-
-
-class ImagingCompressors:
-    class Compressors(Enum):
-        """Enum for compression algorithms a user can select"""
-
-        blosc = Blosc.codec_id
-
-    compressors = [member.value for member in Compressors]
-
-    @staticmethod
-    def get_compressor(compressor_name, **kwargs):
-        """
-        Retrieve a compressor for a given name and optional kwargs.
-        Args:
-            compressor_name (str): Matches one of the names Compressors enum
-            **kwargs (dict): Options to pass into the Compressor
-        Returns:
-            An instantiated compressor class.
-        """
-        if compressor_name == ImagingCompressors.Compressors.blosc.name:
-            return Blosc(**kwargs)
-        else:
-            raise Exception(
-                f"Unknown compressor. Please select one of "
-                f"{ImagingCompressors.compressors}"
             )
