+++ src/aind_data_transfer/writers/__init__.py
@@ -1,0 +1,1 @@
+"""Package for modules to write data."""
