+++ tests/test_compressors.py
@@ -6,12 +6,13 @@ import numpy as np
 from numcodecs import Blosc
 from wavpack_numcodecs import WavPack
 
-from aind_data_transfer.readers import EphysReaders
-from aind_data_transfer.transformations.compressors import (
+from aind_data_transfer.readers.ephys_readers import EphysReaders
+from aind_data_transfer.transformations.ephys_compressors import (
     EphysCompressors,
-    ImagingCompressors,
     VideoCompressor,
 )
+from aind_data_transfer.transformations.imaging_compressors import \
+    ImagingCompressors
 
 TEST_DIR = Path(os.path.dirname(os.path.realpath(__file__)))
 RESOURCES_DIR = TEST_DIR / "resources"
